Move type adjustment logic to MappingElement
